Fix compliation issues (#121)
* Fix errors and warnings connected to lwip2, so that project compiles for NodeMCU v1.0
* Adapt code to change in Neopixelbus
* Add ESPAsyncUDP dependency to travis.
* Add aJson patch and commit hashes to buildscript
* Change paths in buildscript
* Fix build failure with another path using environment variable.
* Unbreak ArduinoJson
* Change WebServer dependencies and NeoPixel method as suggested by Makuna in issue #120

+++ ESP8266HueEmulator/LightService.cpp
@@ -134,7 +134,7 @@ public:
     {
       assert(_wildcard != '/');
       // verify that the URI is reasonable (only contains wildcard at the beginning/end/whole path segments
-      for(int i = 0; i < _uri.length(); i++) {
+      for(size_t i = 0; i < _uri.length(); i++) {
         if (_uri[i] == _wildcard) {
           if (i != 0 && i != _uri.length() - 1 && (_uri[i-1] != '/' || _uri[i+1] != '/')) {
             assert(false);
